persist trades n dividends

+++ spec/jasmine_examples/AccountPersistenceSpec.js
@@ -1,52 +1,61 @@
 describe('Account', function () {
 
-    var account = require('../../src/Account.js');
-    var BigNumber = require('bignumber.js');
-    var make = require('../../src/Trade.js');
-
-    var getAccountMapper = require('../../src/dataMapper/AccountMapper.js');
-
+    var path = require('path');
     var async = require('async');
+    var BigNumber = require('bignumber.js');
+
+    var make = require('../../src/Trade.js');
+    var account = require('../../src/Account.js');
+    var parse = require('../../src/CsvParser.js');
+    var getAccountMapper = require('../../src/dataMapper/AccountMapper.js');
+    var istream;
 
     beforeEach(() => {
         var fs = require('fs');
-        fs.unlinkSync('./stockmon.sqlite');
+        fs.access('./stockmon.sqlite', fs.F_OK, err => {
+            var fileExists = !err;
+            if (fileExists) { fs.unlinkSync('./stockmon.sqlite'); }
+        });
+
+        var csv_path = path.resolve(__dirname, 'datafiles', 'sample_trades.csv');
+        istream = fs.createReadStream(csv_path);
     });
 
-    var trades = [make.makeBuy('2008-07-14', 'HDFCBANK', 10, '1030', '69.01'),
-        make.makeBuy('2009-09-03', 'HDFC', 4, '2705', '181.24'),
-        make.makeBuy('2009-09-13', 'COALIND', 10, '270', '18.09'),
-        make.makeBuy('2009-09-22', 'HDFC', 76, '1190.5', '606.20'),
-        make.makeSale('2009-09-23', 'HDFCBANK', 5, '1607.1', '53.84'),
-        make.makeSale('2009-09-31', 'COALIND', 10, '290', '19.43')
-    ];
     it('can be persisted', function (done) {
+        parse(istream, function (err, results) {
+            expect(err).toBeNull();
+            var inMemAccount, fromDisk;
+            inMemAccount = account.create('Mushu');
+            inMemAccount.register(results.trades);
+            inMemAccount.addDividends(results.dividends);
 
-        var inMemAccount, fromDisk;
-        inMemAccount = account.create('Mushu');
-        inMemAccount.register(trades);
+            var mapper = getAccountMapper();
+            async.waterfall([
+                (cb) => {
+                    mapper.save(inMemAccount, (err, id) => {
+                        cb(null, id);
+                    });
+                },
+                (id, cb) => { var fromDisk = mapper.load(id, (err, account) => { cb(null, account); }); }
+            ],
+                function (err, account) {
+                    if (err) {
+                        done.fail(err.message);
+                    }
+                    expect(account.getId()).toEqual(1);
+                    expect(account.getName()).toEqual('Mushu');
 
-        var mapper = getAccountMapper();
-        async.waterfall([
-            (cb) => {
-                mapper.save(inMemAccount, (err, id) => {
-                    cb(null, id);
+                    var gains = account.getGains();
+                    expect(gains.length).toEqual(5);
+
+                    expect(gains[2].stock).toEqual('SBI');
+                    expect(gains[2].amt.toString()).toEqual('180');
+
+                    expect(gains[4].stock).toEqual('HDFCBANK');
+                    expect(gains[4].qty).toEqual(10);
+
+                    done();
                 });
-            },
-            (id, cb) => { var fromDisk = mapper.load(id, (err, account) => { cb(null, account); }); }
-        ],
-            function (err, account) {
-                if (err) {
-                    done.fail(err.message);
-                }
-                expect(account.getId()).toEqual(1);
-                expect(account.getName()).toEqual('Mushu');
-
-                var holdings = account.getHoldings();
-                expect(holdings[0].stock).toEqual('HDFC');
-                expect(holdings[1].qty).toEqual(5);
-
-                done();
-            });
+        });
     });
 });
